requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
asyncpg==0.30.0
attrs==23.1.0
backcall==0.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2025.1.31
cffi==1.15.1
click==8.1.8
colorama==0.4.6
comm==0.1.3
coverage==7.9.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==1.2.0
fastapi==0.115.6
fastapi-sso==0.17.0
fastjsonschema==2.17.1
fqdn==1.5.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.4
iniconfig==2.1.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.5
jsonpointer==2.4
jsonschema==4.18.1
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jyquickhelper==0.4.220
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdformat==0.7.22
mdurl==0.1.2
mistune==3.0.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
pillow==11.2.1
platformdirs==3.8.1
pluggy==1.5.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.6.1
pycparser==2.21
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.15.1
PyJWT==2.10.1
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.20
pywin32==306; sys_platform == "win32"
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.29.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.10
rsa==4.9
Send2Trash==1.8.2
shiboken6==6.9.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.37
stack-data==0.6.2
starlette==0.41.3
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.13.0
uri-template==1.3.0
uvicorn==0.34.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
